back-end/requirements.txt
aiohttp
aiohttp-cors==0.7.0
async-timeout==3.0.1
attrs
bidict
certifi==2021.5.30
cffi
chardet
cryptography
dnspython==1.16.0
eventlet
greenlet
idna
idna-ssl
multidict
pycparser
pyOpenSSL
python-chess==0.31.4
python-engineio
python-socketio
six
typing-extensions
yarl
